+++ talos.py
@@ -15,7 +15,8 @@ def main():
     #interface.launch_user_interface()
 
     
-    tracker = MediaPipeTracker(args.source)
+    tracker = MediaPipeTracker(args.source, "./config.yaml")
+    #tracker = BasicTracker(args.source, "./config.yaml")
     #director = DiscreteDirector(tracker, "./config.yaml")
     director = ContinuousDirector(tracker, "./config.yaml")
